refactor(main): by default the fluid API is exported
fluid is simply easier to use
