+++ extra.ts
@@ -6,4 +6,3 @@ export {curry, compose} from './src/lib/Utils'
 export {EVENT} from './src/testing/Events'
 export {fromArray} from './src/sources/FromArray'
 export {TestScheduler} from './src/testing/TestScheduler'
-export {Air} from './src/air'
