+++ fp.ts
@@ -1,0 +1,28 @@
+/**
+ * Created by tushar.mathur on 11/10/16.
+ */
+
+export {create} from './src/sources/CreateObservable'
+export {debounce} from './src/operators/Debounce'
+export {delay} from './src/operators/Delay'
+export {filter} from './src/operators/Filter'
+export {forEach} from './src/lib/ForEach'
+export {frames} from './src/sources/Frames'
+export {fromDOM} from './src/sources/FromDOM'
+export {fromPromise} from './src/sources/FromPromise'
+export {interval} from './src/sources/Interval'
+export {join, flatMap} from './src/operators/Join'
+export {map} from './src/operators/Map'
+export {merge} from './src/operators/Merge'
+export {multicast} from './src/operators/Multicast'
+export {Observable} from './src/lib/Observable'
+export {Observer} from './src/lib/Observer'
+export {of} from './src/sources/FromArray'
+export {reduce} from './src/operators/Reduce'
+export {sample} from './src/operators/Sample'
+export {scan} from './src/operators/Scan'
+export {skipRepeats} from './src/operators/SkipRepeats'
+export {slice} from './src/operators/Slice'
+export {SubscriberFunction} from './src/lib/SubscriberFunction'
+export {Subscription} from './src/lib/Subscription'
+export {switchLatest, switchMap} from './src/operators/Switch'
