+++ test/test.Frames.ts
@@ -2,7 +2,7 @@
  * Created by tushar on 28/01/17.
  */
 import {test} from 'ava'
-import {frames} from '../src/main'
+import {frames} from '../fp'
 import {TestScheduler} from '../src/testing/TestScheduler'
 import {toMarble} from '../src/testing/Marble'
 import {scan} from '../src/operators/Scan'
